Increase coverage

+++ api/src/main/java/pcap/api/internal/DefaultTimestamp.java
@@ -25,7 +25,7 @@ public class DefaultTimestamp implements Timestamp {
   }
 
   @Override
-  public int microSecond() {
+  public long microSecond() {
     return microSecond;
   }
